Split directory structure into language and features
