+++ .gitignore
@@ -4,6 +4,7 @@
 *.dll
 *.so
 *.dylib
+.idea/
 
 # Test binary, built with `go test -c`
 *.test
@@ -13,3 +14,4 @@
 
 # Dependency directories (remove the comment below to include it)
 # vendor/
+node_modules
