--- actions/go.mod
@@ -1,5 +1,0 @@
-module github.com/vmware-tanzu/octant-example-plugins/actions
-
-go 1.13
-
-require github.com/vmware-tanzu/octant v0.10.1 // indirect
