+++ golang/actions-api-sample/README.md
@@ -1,0 +1,20 @@
+## Actions API Sample
+
+This example shows how an action can be called to maintain state in a plugin.
+
+Submitting a number to the form saves each subsequent entry.
+
+![action-api-sample-demo](actions-api-sample.gif)
+
+## Installation
+
+Build the plugin:
+
+`go build`
+
+Then move the binary:
+
+`mv actions-api-sample $/HOME/.config/octant/plugins/`
+
+You may need to create this directory if it does not exist.
+
