+++ golang/actions-api-sample/go.mod
@@ -1,0 +1,9 @@
+module github.com/vmware-tanzu/octant-example-plugins/golang/actions-api-sample
+
+go 1.16
+
+require (
+	github.com/magiconair/properties v1.8.1
+	github.com/stretchr/testify v1.6.1
+	github.com/vmware-tanzu/octant v0.17.0
+)
