+++ golang/alert-sample/README.md
@@ -1,0 +1,20 @@
+## Alert Sample
+
+This example shows how to send an alert visible globally above the header bar.
+
+The alert is triggered through an action and is mapped only to the Octant instance that triggered the action.
+
+![alert-sample-demo](alert-sample.gif)
+
+## Installation
+
+Build the plugin:
+
+`go build`
+
+Then move the binary:
+
+`mv alert-sample $/HOME/.config/octant/plugins/`
+
+You may need to create this directory if it does not exist.
+
