+++ golang/alert-sample/go.mod
@@ -1,0 +1,8 @@
+module github.com/vmware-tanzu/octant-example-plugins/golang/alert-sample
+
+go 1.16
+
+require (
+	github.com/vmware-tanzu/octant v0.17.0
+	k8s.io/apimachinery v0.19.3
+)
