+++ golang/extend-gvk-sample/go.mod
@@ -1,0 +1,9 @@
+module github.com/vmware-tanzu/octant-example-plugins/golang/extend-gvk-sample
+
+go 1.16
+
+require (
+	github.com/pkg/errors v0.9.1
+	github.com/vmware-tanzu/octant v0.17.0
+	k8s.io/apimachinery v0.20.4
+)
