+++ golang/module-sample/README.md
@@ -1,0 +1,17 @@
+## Module Sample
+
+This example shows the addition of a new module.
+It adds a new item to the navigation which can have subtrees to show multiple views.
+
+## Installation
+
+Build the plugin:
+
+`go build`
+ 
+Then move the binary:
+
+`mv module-sample $/HOME/.config/octant/plugins/`
+
+You may need to create this directory if it does not exist.
+
