--- sample/go.mod
@@ -1,9 +1,0 @@
-module github.com/vmware-tanzu/octant-example-plugins/sample
-
-go 1.13
-
-require (
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/vmware-tanzu/octant v0.10.1 // indirect
-	k8s.io/apimachinery v0.17.3 // indirect
-)
